message search working & tested. Hi bald yak!
