+++ go/contacts/urls.py
@@ -5,8 +5,8 @@ urlpatterns = patterns('',
     url(r'^$', views.index, name='index'),
     url(r'^groups/$', views.groups, name='groups'),
     # TODO: Is the group_name regex sane?
-    url(r'^groups/(?P<group_key>[\w ]+)/$', views.group, name='group'),
+    url(r'^groups/(?P<group_key>\w+)/$', views.group, name='group'),
     url(r'^people/$', views.people, name='people'),
     url(r'^people/new/$', views.new_person, name='new_person'),
-    url(r'^people/(?P<person_key>\w+)/$', views.person, name='person'),
+    url(r'^people/(?P<person_key>.+)/$', views.person, name='person'),
 )
