+++ go/conversation/templates/generic/includes/message-load-results.html
@@ -24,7 +24,7 @@
         $.ajax({
             url: "{{conversation.get_absolute_url}}results/",
             data: {
-                "query": "{{query}}",
+                "q": "{{query}}",
                 "token": "{{token}}",
                 "batch_id": "{{batch_id}}",
                 "direction": "{{message_direction}}",
